Move tagged_blob macro here from jf-utils
It conceptually belongs here anyways. Also, this fixes the
dependency tree so that `commit` can depend on `tagged_blob` without
depending on `jf-utils` (which would create a cycle). This will
enable tagged base 64 serialization support directly on the
`Commitment` type, obviating the need for many newtype structs.
That will be done in a subsequent change.

During the move, I streamlined the `tagged_blob` macro a bit. I
removed the `TaggedBlob` wrapper type which was only used to add
custom serde impls and to convert to and from for serializing other
types. Instead, serialization now goes directly through `TaggedBase64`,
and I have implemented serialization and deserialization for the
`TaggedBase64` type itself.

Since `TaggedBlob` is gone, I renamed the macro `tagged` for simplicity.